+++ .gitignore
@@ -6,8 +6,6 @@ obj/
 .idea/
 *.sln.iml
 
-Pomnesh/Pomnesh.API
-
 # Visual Studio Code Specific
 .vscode/
